+++ packages/centra-types/src/api/index.ts
@@ -1,0 +1,2 @@
+export { default as ProductResponse } from './Product'
+export * from './Product'
