+++ packages/centra-types/src/index.ts
@@ -1,47 +1,2 @@
-export { default as Address } from './Address'
-export * from './Address'
-
-export { default as BrickAndMortar } from './BrickAndMortar'
-export * from './BrickAndMortar'
-
-export { default as Bundle } from './Bundle'
-export * from './Bundle'
-
-export { default as BundleInfo } from './BundleInfo'
-
-export { default as Category } from './Category'
-
-export { default as Country } from './Country'
-export * from './Country'
-
-export { default as Currency } from './Currency'
-
-export { default as Discounts } from './Discounts'
-
-export { default as Item } from './Item'
-
-export { default as Language } from './Language'
-
-export { default as Location } from './Location'
-
-export { default as LoggedIn } from './LoggedIn'
-
-export { default as Order } from './Order'
-export * from './Order'
-
-export { default as Payment } from './Payment'
-
-export { default as PaymentMethod } from './PaymentMethod'
-
-export { default as PluginFields } from './PluginFields'
-
-export { default as Product } from './Product'
-export * from './Product'
-
-export { default as Response } from './Response'
-export * from './Response'
-
-export { default as Selection } from './Selection'
-export * from './Selection'
-
-export { default as Shipment } from './Shipment'
+export * from './api'
+export * from './models'
