+++ packages/centra-types/src/models/Bundle.ts
@@ -1,16 +1,8 @@
 import Product, { RelatedProduct } from './Product'
-import BundleInfo from './BundleInfo'
+import { BundleInfoExtended } from './BundleInfo'
 
 export default interface Bundle extends Product {
   bundleInfo: BundleInfoExtended
-}
-
-export interface BundleInfoExtended extends BundleInfo {
-  sections: {
-    section: string
-    quantity: number
-    products: number[]
-  }[]
 }
 
 export interface BundleResponse {
