+++ packages/centra-types/src/models/Category.ts
