+++ packages/centra-types/src/models/Filter.ts
@@ -1,0 +1,5 @@
+export default interface Filter {
+  field: string
+  // TODO: add type for filter value
+  values: unknown[]
+}
