+++ packages/centra-types/src/models/Selection.ts
